+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Bump up okapi interfaces for `pieces` (2.0 3.0). Refs UIOR-1212.
 * Invoice line Subscription fields are not populated correctly. Refs UIOR-1220.
 * Opening and editing POs with location-restricted funds (FE). Refs UIOR-1184.
+* Location-restricted funds in POLs. Refs UIOR-1166.
 
 ## [5.0.1](https://github.com/folio-org/ui-orders/tree/v5.0.1) (2023-11-08)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v5.0.0...v5.0.1)